deform2d_double 加了bias

+++ deform3dl/test_model.py
@@ -5,7 +5,7 @@ from torch.autograd import Variable
 from deform3dl.deform_conv3dl_modules import ConvOffset3d
 
 batchsize = 2
-c_in = 1.
+c_in = 1
 c_out = 1
 in_l = in_h = in_w = 3
 kernel_l = kernel_h = kernel_w = 1
@@ -23,7 +23,7 @@ conv = nn.Conv3d(
     bias=False).cuda()
 conv_offset3d = ConvOffset3d(c_in, c_out, (kernel_l, kernel_h, kernel_w), stri, pad, channel_per_group).cuda()
 
-inputs = Variable(torch.randn(batchsize, c_in, in_l, in_h, in_w).type(torch.DoubleTensor).cuda())
+inputs = Variable(torch.randn(batchsize, c_in, in_l, in_h, in_w).cuda())
 offset = conv(inputs)
 output = conv_offset3d(inputs, offset)
 out_0 = torch.rand(*output.size()).type(torch.DoubleTensor).cuda()
